full build of i386 and amd64

+++ download.bash
@@ -1,7 +1,9 @@
 #!/bin/bash
 version=$(cat VERSION)
-wget -O dbrhino-agent_${version}.orig.tar.gz \
-    https://github.com/dbrhino/dbrhino-agent/archive/${version}.tar.gz
+if [[ ! -f dbrhino-agent_${version}.orig.tar.gz ]]; then
+    wget -O dbrhino-agent_${version}.orig.tar.gz \
+        https://github.com/dbrhino/dbrhino-agent/archive/${version}.tar.gz
+fi
 tar xzvf dbrhino-agent_${version}.orig.tar.gz \
     -C dbrhino-agent \
     --strip 1
